requirements.txt
asn1crypto==0.24.0
astroid==2.2.5
attrs==17.4.0
Automat==0.6.0
awscli==1.16.216
blinker==1.4
boto3==1.9.205
botocore==1.12.206
cachetools==3.1.1
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
click==6.7
colorama==0.3.9
configobj==5.0.6
constantly==15.1.0
cryptography==2.1.4
Django==2.0.13
django-cors-headers==3.0.2
docutils==0.14
future==0.17.1
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
httplib2==0.9.2
idna==2.8
jmespath==0.9.4
lazy-object-proxy==1.4.1
pyasn1==0.4.6
pyasn1-modules==0.2.6
pycparser==2.19
PyJWT==1.7.1
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.1
requests==2.22.0
rsa==3.4.2
s3transfer==0.2.1
six==1.12.0
typed-ast==1.4.0
uritemplate==3.0.0
urllib3==1.25.3
wrapt==1.11.2
